pyproject.toml
[tool.poetry]
name = "course-project-7"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.3"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
djangorestframework-simplejwt = "^5.3.1"
django-filter = "^24.3"
django-phonenumber-field = {extras = ["phonenumberslite"], version = "^8.0.0"}
celery = "^5.4.0"
redis = "^5.2.0"
eventlet = "^0.38.0"
django-celery-beat = "^2.7.0"
gevent = "^24.11.1"


[tool.poetry.group.dev.dependencies]
ipython = "^8.29.0"

[tool.poetry.group.lint.dependencies]
isort = "^5.13.2"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
